TeacherDataSystem/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.3
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
pydantic==2.5.0
pydantic-settings==2.1.0
pymupdf==1.24.0
Pillow==10.2.0
reportlab==4.0.9
pdf2image==1.17.0

